pyproject.toml
[project]
name = "scholar-inbox-slack-bot"
version = "0.1.0"
description = "Automated paper summarization bot for Scholar Inbox with Slack integration"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.72.0",
    "apscheduler>=3.11.1",
    "arxiv>=2.3.0",
    "beautifulsoup4>=4.14.2",
    "google-generativeai>=0.8.5",
    "lxml>=6.0.2",
    "openai>=2.6.1",
    "pillow>=12.0.0",
    "playwright>=1.55.0",
    "pydantic>=2.12.3",
    "pypdf>=6.1.3",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "slack-sdk>=3.37.0",
]

[project.scripts]
scholar-bot = "src.main:main"
